requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.1
bson==0.5.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
dnspython==1.16.0
ecdsa==0.19.0
fastapi==0.115.5
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.16.1
fsspec==2024.9.0
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
jwt==1.3.1
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai==1.51.0
openai-whisper==20240930
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pymongo==3.11.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.20
regex==2024.9.11
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
